Massive update which introduces a scheduler in the Client, all debug prints need to be removed still
