+++ binsync/common/ui/panel_tabs/functions_table.py
@@ -12,7 +12,7 @@ from binsync.common.ui.qt_objects import (
 )
 from binsync.common.ui.utils import QNumericItem, friendly_datetime
 from binsync.data import Function
-from binsync.state import State
+from binsync.core.state import State
 
 l = logging.getLogger(__name__)
 
@@ -112,7 +112,7 @@ class QFunctionTable(QTableWidget):
 
         # first check if any functions are unknown to the table
         for user in self.controller.users():
-            state = self.controller.client.get_state(user=user.name, readonly=True)
+            state = self.controller.client.get_state(user=user.name)
             user_funcs: Dict[int, Function] = state.functions
 
             for func_addr, sync_func in user_funcs.items():
@@ -134,8 +134,8 @@ class QFunctionTable(QTableWidget):
         self.items = [QFunctionItem(*row) for row in known_funcs.values()]
 
     def _get_valid_users_for_func(self, func_addr):
-        for user in self.controller.users():
-            user_state: State = self.controller.client.get_state(user=user.name, readonly=True)
+        for user in self.controller.users(priority=1):
+            user_state: State = self.controller.client.get_state(user=user.name, priority=1)
             user_func = user_state.get_function(func_addr)
 
             # function must be changed by this user
