+++ binsync/core/errors.py
